Switched off upower and busctl as they can't do what I need properly.
udevadm should monitor for events specifics to the power_supply area and nothing else.

systemd-ac-power will check all the power adapters and if one of them is online it'll say yes.
if none are then its a no.

These allow a super simple wrapper that will work reliably and faster.